swaggerCodegen/server/test-requirements.txt
flask_testing==0.6.1
coverage>=4.0.3
nose>=1.3.7
pluggy>=0.3.1
py>=1.4.31
randomize>=0.13
pandas >= 0.20.3
numpy >= 1.14.0
